# Task2/e-transport_back/swagger-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/user": {
            "get": {
                "operationId": "UserController_findAllUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all users returned successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "user"
                ]
            },
            "post": {
                "operationId": "UserController_createUser",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully."
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "user"
                ]
            }
        },
        "/user/{id}": {
            "put": {
                "operationId": "UserController_updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully."
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "user"
                ]
            },
            "delete": {
                "operationId": "UserController_deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully."
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "user"
                ]
            },
            "get": {
                "operationId": "UserController_getUserDetails",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details returned successfully."
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "user"
                ]
            }
        },
        "/rental": {
            "get": {
                "operationId": "RentalController_findAllRental",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all rentals returned successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental"
                ]
            },
            "post": {
                "operationId": "RentalController_createRental",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RentalDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental created successfully."
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental"
                ]
            }
        },
        "/rental/{id}": {
            "put": {
                "operationId": "RentalController_updateRental",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRentalDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental updated successfully."
                    },
                    "404": {
                        "description": "Rental not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental"
                ]
            },
            "delete": {
                "operationId": "RentalController_deleteRental",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental deleted successfully."
                    },
                    "404": {
                        "description": "Rental not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental"
                ]
            }
        },
...
        "/vehicle": {
            "get": {
                "operationId": "VehicleController_findAllVehicle",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all vehicles returned successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            },
            "post": {
                "operationId": "VehicleController_createVehicle",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created successfully."
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            }
        },
        "/vehicle/{id}": {
            "put": {
                "operationId": "VehicleController_updateVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated successfully."
                    },
                    "404": {
                        "description": "Vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            },
            "delete": {
                "operationId": "VehicleController_deleteVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle deleted successfully."
                    },
                    "404": {
                        "description": "Vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            }
        },
...
    },
    "info": {
        "title": "e-transport",
        "description": "The e-transport API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "UserDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "bonusAccount": {
                        "type": "string"
                    },
                    "notification": {
                        "type": "boolean"
                    },
                    "photo": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "USER",
                            "MODERATOR",
                            "TECHNICIAN"
                        ]
                    }
                },
                "required": [
                    "email",
                    "name",
                    "password",
                    "phoneNumber",
                    "bonusAccount",
                    "notification",
                    "photo",
                    "role"
                ]
            },
            "UpdateUserDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "bonusAccount": {
                        "type": "string"
                    },
                    "photo": {
                        "type": "string"
                    },
                    "notification": {
                        "type": "boolean"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "USER",
                            "MODERATOR",
                            "TECHNICIAN"
                        ]
                    }
                },
                "required": [
                    "role"
                ]
            },
            "RentalDto": {
                "type": "object",
                "properties": {
                    "isActive": {
                        "type": "boolean"
                    },
                    "dateRented": {
                        "type": "string"
                    },
                    "dateReturned": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "distance": {
                        "type": "number"
                    },
                    "avgSpeed": {
                        "type": "number"
                    },
                    "maxSpeed": {
                        "type": "number"
                    },
                    "energyConsumed": {
                        "type": "number"
                    }
                },
                "required": [
                    "isActive",
                    "dateRented",
                    "dateReturned",
                    "userId",
                    "distance",
                    "avgSpeed",
                    "maxSpeed",
                    "energyConsumed"
                ]
            },
            "UpdateRentalDto": {
                "type": "object",
                "properties": {
                    "isActive": {
                        "type": "boolean"
                    },
                    "dateRented": {
                        "type": "string"
                    },
                    "dateReturned": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "distance": {
                        "type": "number"
                    },
                    "avgSpeed": {
                        "type": "number"
                    },
                    "maxSpeed": {
                        "type": "number"
                    },
                    "energyConsumed": {
                        "type": "number"
                    }
                }
            },
...
            "VehicleDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "FREE",
                            "INUSE",
                            "NOTAVAILABLE",
                            "BROKEN",
                            "REPAIR"
                        ]
                    },
                    "runnedDistance": {
                        "type": "number"
                    },
                    "releaseDate": {
                        "type": "string"
                    },
                    "currentLocation": {
                        "type": "string"
                    }
                },
                "required": [
                    "status",
                    "runnedDistance",
                    "releaseDate",
                    "currentLocation"
                ]
            },
            "UpdateVehicleDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "FREE",
                            "INUSE",
                            "NOTAVAILABLE",
                            "BROKEN",
                            "REPAIR"
                        ]
                    },
                    "runnedDistance": {
                        "type": "number"
                    },
                    "releaseDate": {
                        "type": "string"
                    },
                    "currentLocation": {
                        "type": "string"
                    }
                },
                "required": [
                    "status"
                ]
            },
...
        }
    }
}